+++ tests/flex_wrap_test.rs
@@ -1546,7 +1546,7 @@ fn test_wrapped_column_max_height_flex() {
 
     let mut root_child0 = Node::new_with_config(&mut config);
     root_child0.set_flex_grow(1_f32);
-    root_child0.set_flex_shrink(1_f32 as f32);
+    root_child0.set_flex_shrink(1_f32);
     root_child0.set_flex_basis(StyleUnit::Percent((0 as f32).into()));
     root_child0.set_width(StyleUnit::Point(100_f32.into()));
     root_child0.set_min_width(StyleUnit::Auto);
@@ -1557,7 +1557,7 @@ fn test_wrapped_column_max_height_flex() {
 
     let mut root_child1 = Node::new_with_config(&mut config);
     root_child1.set_flex_grow(1_f32);
-    root_child1.set_flex_shrink(1_f32 as f32);
+    root_child1.set_flex_shrink(1_f32);
     root_child1.set_flex_basis(StyleUnit::Percent((0 as f32).into()));
     root_child1.set_margin(Edge::Left, StyleUnit::Point(20_f32.into()));
     root_child1.set_margin(Edge::Top, StyleUnit::Point(20_f32.into()));
