+++ tests/min_max_dimension_test.rs
@@ -365,7 +365,7 @@ fn test_flex_grow_to_min() {
 
     let mut root_child0 = Node::new_with_config(&mut config);
     root_child0.set_flex_grow(1_f32);
-    root_child0.set_flex_shrink(1_f32 as f32);
+    root_child0.set_flex_shrink(1_f32);
     root_child0.set_min_width(StyleUnit::Auto);
     root_child0.set_min_height(StyleUnit::Auto);
     root.insert_child(&mut root_child0, 0);
@@ -998,7 +998,7 @@ fn test_flex_grow_within_constrained_max_row() {
     root.insert_child(&mut root_child0, 0);
 
     let mut root_child0_child0 = Node::new_with_config(&mut config);
-    root_child0_child0.set_flex_shrink(1_f32 as f32);
+    root_child0_child0.set_flex_shrink(1_f32);
     root_child0_child0.set_flex_basis(StyleUnit::Point(100_f32.into()));
     root_child0_child0.set_min_width(StyleUnit::Auto);
     root_child0_child0.set_min_height(StyleUnit::Auto);
@@ -1063,7 +1063,7 @@ fn test_flex_grow_within_constrained_max_column() {
     root.set_max_height(StyleUnit::Point(100_f32.into()));
 
     let mut root_child0 = Node::new_with_config(&mut config);
-    root_child0.set_flex_shrink(1_f32 as f32);
+    root_child0.set_flex_shrink(1_f32);
     root_child0.set_flex_basis(StyleUnit::Point(100_f32.into()));
     root_child0.set_min_width(StyleUnit::Auto);
     root_child0.set_min_height(StyleUnit::Auto);
